add: employers can update company details, add: remove refresh token on logout
