+++ task1/job-board/src/screens/profile/Profile.css
@@ -1,0 +1,6 @@
+.category-input{
+    @apply w-full h-fit pb-4 flex flex-col p-2 border border-slate-400 rounded
+}
+/*.category-input > h3{*/
+/*    @apply text-center text-stone-700 font-semibold*/
+/*}*/
